+++ microservice/microserviceTests/microservice/SimpleMicroservice.cs
@@ -29,9 +29,9 @@ namespace microserviceTests.microservice
    [StorageObject("simple")]
    public class SimpleStorageObject : MongoStorageObject
    {
-
+      
    }
-
+   
    [Microservice("simple", UseLegacySerialization = true)]
    public class SimpleMicroservice : Microservice
    {
@@ -75,7 +75,7 @@ namespace microserviceTests.microservice
          await Task.Delay(ms);
          return ms;
       }
-
+      
       [ClientCallable]
       public async Task<string> DelayThenGetEmail(int ms, long dbid)
       {
@@ -195,6 +195,12 @@ namespace microserviceTests.microservice
       {
          var res = await Services.Leaderboards.GetPlayerLeaderboards(dbid);
          return res.lbs.Count;
+      }
+      
+      [ClientCallable]
+      public async Task RemovePlayerEntry(string leaderboardId, long dbid)
+      {
+	      await Services.Leaderboards.RemovePlayerEntry(leaderboardId, dbid);
       }
 
       [ClientCallable]