+++ microservice/microservice/dbmicroservice/BeamableMicroService.cs
@@ -92,6 +92,7 @@ namespace Beamable.Server
       private const int EXIT_CODE_FAILED_CUSTOM_INITIALIZATION_HOOK = 110;
       private const int HTTP_STATUS_GONE = 410;
       private const int ShutdownLimitSeconds = 5;
+      private const int ShutdownMinCycleTimeMilliseconds = 100;
       private Promise<Unit> _serviceInitialized = new Promise<Unit>();
       private IConnection _connection;
       private readonly IConnectionProvider _connectionProvider;
@@ -152,7 +153,7 @@ namespace Beamable.Server
       };
 
       private int _connectionAttempt = 0;
-      
+
       /// <summary>
       /// We need to guarantee <see cref="ResolveCustomInitializationHook"/> only gets run once when we <see cref="SetupWebsocket"/>.
       /// </summary>
@@ -213,12 +214,15 @@ namespace Beamable.Server
          _reflectionCache.GenerateReflectionCache(relevantAssemblyNames);
 
          _socketRequesterContext = new SocketRequesterContext(GetWebsocketPromise);
-         _requester = new MicroserviceRequester(_args, null, _socketRequesterContext);
+         _requester = new MicroserviceRequester(_args, null, _socketRequesterContext, false);
          _mongoSerializationService = new MongoSerializationService();
          _storageObjectConnectionProviderService = new StorageObjectConnectionProvider(_args, _requester);
 
          _contentService = new ContentService(_requester, _socketRequesterContext, _contentResolver, _reflectionCache);
          ContentApi.Instance.CompleteSuccess(_contentService);
+
+         _serviceShutdownTokenSource = new CancellationTokenSource();
+         (_socketDaemen, _socketRequesterContext.Daemon) = MicroserviceAuthenticationDaemon.Start(_args, _requester, _serviceShutdownTokenSource);
 
          InitServices();
 
@@ -234,9 +238,6 @@ namespace Beamable.Server
          // Connect and Run
          _webSocketPromise = AttemptConnection();
          var socket = await _webSocketPromise;
-
-         _serviceShutdownTokenSource = new CancellationTokenSource();
-         _socketDaemen = MicroserviceAuthenticationDaemon.Start(_args, _requester, _socketRequesterContext, _serviceShutdownTokenSource);
 
          var setupWebsocketTask = SetupWebsocket(socket);
          setupWebsocketTask.Wait();
@@ -291,8 +292,17 @@ namespace Beamable.Server
             var secondsLeft = millisecondsLeft / 1000;
             Log.Debug("Waiting up to {shutdownTimeLimit} seconds for tasks to complete.", secondsLeft);
             var pendingTasks = _runningTaskTable.Values.ToArray();
+            var startedWaitingAt = sw.ElapsedMilliseconds;
             Task.WaitAll(pendingTasks, TimeSpan.FromMilliseconds(millisecondsLeft));
-
+            /* we need to wait a minimum number of moments in this loop so we don't exhaust the task cycle.
+             What can happen is that all the tasks are DONE, so, the WaitAll completes. However, the task hasn't removed itself from the _runningTaskTable yet,
+             because its continuation hasn't been executed. And then, if that happens, we enter a scenario where this method just loops and loops, and since the Task.WaitAll
+             is always complete, the other continuation takes a long time to get scheduled.
+             */
+            var stoppedWaitingAt = sw.ElapsedMilliseconds;
+            var waitedForMilliseconds = stoppedWaitingAt - startedWaitingAt;
+            var requiredWaitTimeLeft = Math.Max(0, ShutdownMinCycleTimeMilliseconds - waitedForMilliseconds);
+            await Task.Delay(TimeSpan.FromMilliseconds(requiredWaitTimeLeft));
          }
          if (_runningTaskTable.Count > 0)
          {
@@ -305,7 +315,7 @@ namespace Beamable.Server
          }
 
          // stop the daemon from trying to re-authenticate
-         MicroserviceAuthenticationDaemon.KillAuthThread(_serviceShutdownTokenSource);
+         _socketRequesterContext.Daemon.KillAuthThread();
          await _socketDaemen;
 
          // close the connection itself
@@ -352,7 +362,7 @@ namespace Beamable.Server
 
          try
          {
-            MicroserviceAuthenticationDaemon.WakeAuthThread(ref _socketRequesterContext.AuthorizationCounter);
+	         _socketRequesterContext.Daemon.WakeAuthThread();
             await _requester.WaitForAuthorization();
 
             // Custom Initialization hook for C#MS --- will terminate MS user-code throws.
@@ -362,7 +372,7 @@ namespace Beamable.Server
                await ResolveCustomInitializationHook();
                _ranCustomUserInitializationHooks = true;
             }
-            
+
 
             await ProvideService(QualifiedName);
 
@@ -378,7 +388,7 @@ namespace Beamable.Server
 
       }
 
-      
+
 
       /// <summary>
       /// Handles custom initialization hooks. Makes the following assumptions:
@@ -585,9 +595,10 @@ namespace Beamable.Server
                .AddScoped(MicroserviceType)
                .AddSingleton<IDependencyProvider>(provider => new MicrosoftServiceProviderWrapper(provider))
                .AddSingleton(_args)
+               .AddSingleton(_socketRequesterContext.Daemon)
                .AddSingleton<IRealmInfo>(_args)
                .AddSingleton<SocketRequesterContext>(_ => _socketRequesterContext)
-               .AddTransient<IBeamableRequester, MicroserviceRequester>()
+               .AddTransient<IBeamableRequester, MicroserviceRequester>((provider) => new MicroserviceRequester(_args, provider.GetService<RequestContext>(), _socketRequesterContext, true))
                .AddTransient<IUserContext>(provider => provider.GetService<RequestContext>())
                .AddTransient<IMicroserviceAuthApi, ServerAuthApi>()
                .AddTransient<IMicroserviceStatsApi, MicroserviceStatsApi>()
@@ -815,7 +826,7 @@ namespace Beamable.Server
 
       private IBeamableRequester GenerateRequester(RequestContext ctx)
       {
-         return new MicroserviceRequester(_args, ctx, _socketRequesterContext);
+         return new MicroserviceRequester(_args, ctx, _socketRequesterContext, true);
       }
 
       private IBeamableServices ExtractSdks(IServiceProvider provider)