+++ client/Packages/com.beamable.server/CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `RemovePlayerEntry` for leaderboards API which allows to remove given player from the leaderboard
 
 ## [1.2.5]
 ### Fixed