+++ client/Packages/com.beamable.server/CHANGELOG.md
@@ -5,9 +5,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `RemovePlayerEntry` for leaderboards API which allows to remove given player from the leaderboard
+
+## [1.2.5]
 ### Fixed
 - Failed promises no longer log exception info after an exception handler is registered on the same execution cycle.
 - "Connection is closed" log exception no longer prints incorrectly.
+- Requests no longer attempt to send while authorization process is happening.
 
 ## [1.2.4]
 ### Added